Added resource not found exception and response dto.
