+++ src/main/java/com/MTAPizza/Sympoll/pollmanagementservice/dto/error/GeneralPollErrorResponse.java
@@ -1,5 +1,4 @@
 package com.MTAPizza.Sympoll.pollmanagementservice.dto.error;
 
-public record IllegalArgumentError(String message) {
-
+public record GeneralPollErrorResponse(String message) {
 }
