+++ src/main/java/com/MTAPizza/Sympoll/pollmanagementservice/dto/error/IllegalArgumentResponse.java
@@ -1,0 +1,5 @@
+package com.MTAPizza.Sympoll.pollmanagementservice.dto.error;
+
+public record IllegalArgumentResponse(String message) {
+
+}
