+++ src/main/java/com/MTAPizza/Sympoll/pollmanagementservice/dto/error/JsonParserErrorResponse.java
@@ -1,4 +1,4 @@
 package com.MTAPizza.Sympoll.pollmanagementservice.dto.error;
 
-public record GeneralPollError(String message) {
+public record JsonParserErrorResponse(String message) {
 }
