+++ src/main/java/com/MTAPizza/Sympoll/pollmanagementservice/exception/ResourceNotFoundException.java
@@ -1,0 +1,7 @@
+package com.MTAPizza.Sympoll.pollmanagementservice.exception;
+
+public class ResourceNotFoundException extends RuntimeException{
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
